pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sql-debugger-env"
version = "1.0.0"
description = "OpenEnv RL environment for SQL query debugging and optimisation"
readme = "README.md"
license = { text = "BSD-3-Clause" }
requires-python = ">=3.10"

keywords = [
    "openenv",
    "reinforcement-learning",
    "sql",
    "debugging",
    "llm-training",
    "rl-environment",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Client-side dependencies (lightweight — no DB engine needed)
dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=11.0",
    "requests>=2.28.0",
]
dev = [
    "pytest>=7.0",
    "httpx>=0.24.0",
    "openai>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sql_debugger_env*"]

[tool.setuptools.package-data]
"*" = ["openenv.yaml", "README.md"]

# ── openenv CLI metadata ──────────────────────────────────────────────────────
[tool.openenv]
name          = "sql-debugger-env"
version       = "1.0.0"
description   = "RL environment for SQL query debugging and optimisation"
tags          = ["openenv", "sql", "debugging", "real-world", "code-repair"]
server_dir    = "server"
app_module    = "server.app:app"
